First commit of Mersine Primes
I kinda like this one because it is defiantly more readable than some of my previous ones. 

Finds all of the mersine primes in the range of 2^1 to 2^10. 
Can be improved by making the isPrime method more efficient

I can skip every even number since that is divisible by 2
or I can skip by every n prime that is divisible by n 

Not sure if the above statement is true but it works for 5 and 2 so whatever
Maybe I can make a more optimized prime checker in the future