Create munchhausenNumbers.py
I think there is possibly a way to make this simpler using a sum() method but can't think of anything right now. 
possibly if it is iterable making it an int and then powering it to itself 
not sure though right now. 

Not that it matters since there is only one