Create magicSquare.py
Im going to do a side step and just tell if it is a magic square or not does it's vertical, horizontal, and diagonals all add up to the same number. This also works for N sized magic squares which is cool. Probably can be improved using the sum() function. I just don't know how to manipulate the numbers into working that way. It would be nice if I can make something that can solve a partially solved magic square. 

I might be able to make a solver assuming it knows what the goal is 🤞