Create luckyNumbers.py
This was one of the hardest ones to figure out in my head for some reason. I was stuck trying to find a way to make it find the nth number of a list and delete it. It was very bad it took many pieces of paper, and a couple days which was a first for just finding an integer sequence. I had an epiphany when watching Darek Banas Python in 1 video video. Using the list functions
https://www.youtube.com/watch?v=H1elmMBnykA&t=2939s
and 
https://stackoverflow.com/questions/9027862/what-does-listxy-do
These links really guided my way in finding the solution.