Create nepalGeometry.py
I moved from Texas to Indiana so I haven't been updating my projects even though I have been working on them. But this is a major project that is going to help with the new nepal flag

A lot of stuff is going on in this. Mainly it is to create Points, Lines, and Circles
Then using other functions to figure out where the Intersection points are on lines and lines, circles and circles, and lines and circles. Right now the hardest one that I havn't yet figured out is the lines and circle intersections.